+++ src/main/java/com/snackgame/server/applegame/business/domain/exception/AppleGameException.java
@@ -1,4 +1,4 @@
-package com.snackgame.server.applegame.domain.exception;
+package com.snackgame.server.applegame.business.domain.exception;
 
 public abstract class AppleGameException extends RuntimeException {
