+++ src/test/java/com/snackgame/server/applegame/domain/AppleTest.java
@@ -8,7 +8,8 @@ import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
 
-import com.snackgame.server.applegame.domain.exception.AppleNumberRangeException;
+import com.snackgame.server.applegame.business.domain.Apple;
+import com.snackgame.server.applegame.business.domain.exception.AppleNumberRangeException;
 
 @SuppressWarnings("NonAsciiCharacters")
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
